ヘッダーのカスタマイズに関してのテストを追加 #24
